feat: 스튜디오 생성, 수정 응답 포맷 수정
