+++ src/main/kotlin/kr/kro/dearmoment/studio/application/port/input/GetStudioUseCase.kt
@@ -1,7 +1,7 @@
 package kr.kro.dearmoment.studio.application.port.input
 
-import kr.kro.dearmoment.studio.application.dto.response.StudioResponse
+import kr.kro.dearmoment.studio.application.dto.response.GetStudioResponse
 
 interface GetStudioUseCase {
-    fun getStudio(studioId: Long): StudioResponse
+    fun getStudio(studioId: Long): GetStudioResponse
 }
