+++ src/test/kotlin/kr/kro/dearmoment/studio/application/service/StudioServiceTest.kt
@@ -8,10 +8,11 @@ import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
 import io.mockk.verify
+import kr.kro.dearmoment.common.fixture.studioFixture
 import kr.kro.dearmoment.studio.application.command.ModifyStudioCommand
 import kr.kro.dearmoment.studio.application.command.RegisterStudioCommand
 import kr.kro.dearmoment.studio.application.command.StudioPartnerShopCommand
-import kr.kro.dearmoment.studio.application.dto.response.StudioResponse
+import kr.kro.dearmoment.studio.application.dto.response.GetStudioResponse
 import kr.kro.dearmoment.studio.application.port.output.DeleteStudioPort
 import kr.kro.dearmoment.studio.application.port.output.GetStudioPort
 import kr.kro.dearmoment.studio.application.port.output.SaveStudioPort
@@ -75,21 +76,8 @@ class StudioServiceTest : DescribeSpec({
         }
 
         context("studioId를 전달하면") {
-            val studio =
-                Studio(
-                    userId = 1L,
-                    name = "스튜디오",
-                    contact = "010-1234-5678",
-                    studioIntro = "스튜디오 소개글",
-                    artistsIntro = "작가 소개글",
-                    instagramUrl = "인스타 url",
-                    kakaoChannelUrl = "카카오톡 채널 url",
-                    reservationNotice = "",
-                    cancellationPolicy = "",
-                    status = StudioStatus.ACTIVE,
-                )
-
-            val expected = StudioResponse.from(studio)
+            val studio = studioFixture()
+            val expected = GetStudioResponse.from(studio)
 
             every { getStudioPort.findById(studio.id) } returns studio
             it("studioId에 해당하는 스튜디오를 반환한다.") {
